Cleanup. Add actions file. Move stuff around.
