--- src/actions/filter.js
@@ -1,6 +1,0 @@
-export function doFilterA(data) {
-    return {
-      type: 'DUMMY',
-      params: {a: 1, b: 2}  // params from input even, for example
-    }
-}
